+++ packages/chameleon-tool/commanders/dev/index.js
@@ -5,6 +5,7 @@ exports.desc = 'start dev mode';
 /* istanbul ignore next */
 exports.register = function (commander) {
   commander
+    .option('-n, --nopreview ', "don't auto open preview")
     .action(function (...args) {
       cml.utils.checkProjectConfig();
       /* eslint-disable */